move remaining AuthMiddleware to Middleware
