+++ zio-http/src/main/scala/zhttp/http/Middleware.scala
@@ -4,8 +4,12 @@ import io.netty.handler.codec.http.HttpHeaderNames
 import zhttp.http.CORS.DefaultCORSConfig
 import zhttp.http.Headers.BasicSchemeName
 import zio.clock.Clock
+import zio.console.Console
 import zio.duration.Duration
-import zio.{UIO, ZIO}
+import zio.{UIO, ZIO, clock, console}
+
+import java.io.IOException
+import java.util.UUID
 
 /**
  * Middlewares are essentially transformations that one can apply on any Http to produce a new one. They can modify
@@ -120,6 +124,147 @@ object Middleware {
   def makeZIO[A]: PartialMakeZIO[A] = new PartialMakeZIO[A](())
 
   def succeed[B](b: B): Middleware[Any, Nothing, Nothing, Any, Any, B] = fromHttp(Http.succeed(b))
+
+  /**
+   * Sets cookie in response headers
+   */
+  def addCookie(cookie: Cookie): HttpMiddleware[Any, Nothing] =
+    addHeader(Headers.setCookie(cookie))
+
+  /**
+   * Adds the provided header and value to the response
+   */
+  def addHeader(name: String, value: String): HttpMiddleware[Any, Nothing] =
+    patch((_, _) => Patch.addHeader(name, value))
+
+  /**
+   * Adds the provided header to the response
+   */
+  def addHeader(header: Headers): HttpMiddleware[Any, Nothing] =
+    patch((_, _) => Patch.addHeader(header))
+
+  /**
+   * Adds the provided list of headers to the response
+   */
+  def addHeaders(headers: Headers): HttpMiddleware[Any, Nothing] =
+    patch((_, _) => Patch.addHeader(headers))
+
+  def addCookieM[R, E](cookie: ZIO[R, E, Cookie]): HttpMiddleware[R, E] =
+    patchZIO((_, _) => cookie.mapBoth(Option(_), c => Patch.addHeader(Headers.setCookie(c))))
+
+  /**
+   * CSRF middlewares : To prevent Cross-site request forgery attacks. This middleware is modeled after the double
+   * submit cookie pattern.
+   * @see
+   *   [[Middleware#csrfGenerate]] - Sets cookie with CSRF token
+   * @see
+   *   [[Middleware#csrfValidate]] - Validate token value in request headers against value in cookies
+   * @see
+   *   https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#double-submit-cookie
+   */
+
+  def csrfGenerate[R, E](
+    tokenName: String = "x-csrf-token",
+    tokenGen: ZIO[R, Nothing, String] = UIO(UUID.randomUUID.toString),
+  ): HttpMiddleware[R, E] =
+    addCookieM(tokenGen.map(Cookie(tokenName, _)))
+
+  def csrfValidate(tokenName: String = "x-csrf-token"): HttpMiddleware[Any, Nothing] = {
+    Middleware.whenHeader(
+      headers => {
+        (headers.getHeaderValue(tokenName), headers.getCookieValue(tokenName)) match {
+          case (Some(headerValue), Some(cookieValue)) => headerValue != cookieValue
+          case _                                      => true
+        }
+      },
+      Middleware.fromHttp(Http.status(Status.FORBIDDEN)),
+    )
+  }
+
+  /**
+   * Add log status, method, url and time taken from req to res
+   */
+  def debug: HttpMiddleware[Console with Clock, IOException] =
+    makeResponseZIO((method, url, _) => zio.clock.nanoTime.map(start => (method, url, start))) {
+      case (status, _, (method, url, start)) =>
+        for {
+          end <- clock.nanoTime
+          _   <- console
+            .putStrLn(s"${status.asJava.code()} ${method} ${url.asString} ${(end - start) / 1000000}ms")
+            .mapError(Option(_))
+        } yield Patch.empty
+    }
+
+  /**
+   * Creates a new middleware using transformation functions
+   */
+  def makeResponse[S](req: (Method, URL, Headers) => S): PartialResponseMake[S] = PartialResponseMake(req)
+
+  /**
+   * Creates a new middleware using effectful transformation functions
+   */
+  def makeResponseZIO[R, E, S](req: (Method, URL, Headers) => ZIO[R, Option[E], S]): PartialResponseMakeZIO[R, E, S] =
+    PartialResponseMakeZIO(req)
+
+  /**
+   * Creates a middleware that produces a Patch for the Response
+   */
+  def patch[R, E](f: (Status, Headers) => Patch): HttpMiddleware[R, E] =
+    makeResponse((_, _, _) => ())((status, headers, _) => f(status, headers))
+
+  /**
+   * Creates a middleware that produces a Patch for the Response effectfully.
+   */
+  def patchZIO[R, E](f: (Status, Headers) => ZIO[R, Option[E], Patch]): HttpMiddleware[R, E] =
+    makeResponseZIO((_, _, _) => ZIO.unit)((status, headers, _) => f(status, headers))
+
+  /**
+   * Removes the header by name
+   */
+  def removeHeader(name: String): HttpMiddleware[Any, Nothing] =
+    patch((_, _) => Patch.removeHeaders(List(name)))
+
+  /**
+   * Runs the effect after the response is produced
+   */
+  def runAfter[R, E](effect: ZIO[R, E, Any]): HttpMiddleware[R, E] =
+    patchZIO((_, _) => effect.mapBoth(Option(_), _ => Patch.empty))
+
+  /**
+   * Runs the effect before the request is passed on to the HttpApp on which the middleware is applied.
+   */
+  def runBefore[R, E](effect: ZIO[R, E, Any]): HttpMiddleware[R, E] =
+    makeResponseZIO((_, _, _) => effect.mapError(Option(_)).unit)((_, _, _) => UIO(Patch.empty))
+
+  /**
+   * Creates a new middleware that always sets the response status to the provided value
+   */
+  def status(status: Status): HttpMiddleware[Any, Nothing] = patch((_, _) => Patch.setStatus(status))
+
+  /**
+   * Applies the middleware on an HttpApp
+   */
+
+  final case class PartialResponseMake[S](req: (Method, URL, Headers) => S) extends AnyVal {
+    def apply(res: (Status, Headers, S) => Patch): HttpMiddleware[Any, Nothing] = {
+      Middleware.intercept[Request, Response](
+        incoming = request => req(request.method, request.url, request.getHeaders),
+      )(
+        outgoing = (response, state) => res(response.status, response.getHeaders, state)(response),
+      )
+    }
+  }
+
+  final case class PartialResponseMakeZIO[R, E, S](req: (Method, URL, Headers) => ZIO[R, Option[E], S]) extends AnyVal {
+    def apply[R1 <: R, E1 >: E](res: (Status, Headers, S) => ZIO[R1, Option[E1], Patch]): HttpMiddleware[R1, E1] =
+      Middleware
+        .interceptZIO[Request, Response]
+        .apply[R1, E1, S, Response](
+          incoming = request => req(request.method, request.url, request.getHeaders),
+        )(
+          outgoing = (response, state) => res(response.status, response.getHeaders, state).map(patch => patch(response)),
+        )
+  }
 
   /**
    * Logical operator to decide which middleware to select based on the header
@@ -256,7 +401,7 @@ object Middleware {
             ),
           )
         case (_, Some(origin), _) if allowCORS(origin, method)                                                    =>
-          HttpMiddleware.addHeader(corsHeaders(origin, method, isPreflight = false))
+          Middleware.addHeader(corsHeaders(origin, method, isPreflight = false))
         case _ => Middleware.identity
       }
     })
