+++ zio-http/src/test/scala/zhttp/middleware/HttpMiddlewareSpec.scala
@@ -1,6 +1,6 @@
 package zhttp.middleware
 
-import zhttp.http.Middleware.{cors, ifThenElse, ifThenElseZIO, timeout, whenZIO}
+import zhttp.http.Middleware._
 import zhttp.http._
 import zhttp.internal.HttpAppTestExtensions
 import zio._
@@ -14,14 +14,13 @@ object HttpMiddlewareSpec extends DefaultRunnableSpec with HttpAppTestExtensions
   private val app: HttpApp[Any with Clock, Nothing] = Http.collectZIO[Request] { case Method.GET -> !! / "health" =>
     UIO(Response.ok).delay(1 second)
   }
-  private val midA                                  = HttpMiddleware.addHeader("X-Custom", "A")
-  private val midB                                  = HttpMiddleware.addHeader("X-Custom", "B")
+  private val midA                                  = Middleware.addHeader("X-Custom", "A")
+  private val midB                                  = Middleware.addHeader("X-Custom", "B")
   private val basicHS                               = Headers.basicAuthorizationHeader("user", "resu")
   private val basicHF                               = Headers.basicAuthorizationHeader("user", "user")
   private val basicAuthM                            = Middleware.basicAuth { case (u, p) => p.toString.reverse == u }
 
   def spec = suite("HttpMiddleware") {
-    import HttpMiddleware._
 
     suite("debug") {
       testM("log status method url and time") {
